+++ tests/test_sensor_wiring.py
@@ -45,7 +45,8 @@ def _enum_like(*members):
 
 
 sys.modules["homeassistant.components.sensor"].SensorStateClass = _enum_like("measurement")
-sys.modules["homeassistant.components.sensor"].SensorDeviceClass = _enum_like("timestamp")
+sys.modules["homeassistant.components.sensor"].SensorDeviceClass = _enum_like(
+    "timestamp", "duration")
 
 
 def _load(name, path):